+++ src/TestSuite/TestCase.php
@@ -82,14 +82,16 @@ abstract class TestCase extends CakeTestCase
     /**
      * This method is called after the last test of this test class is run
      * @return void
+     * @codeCoverageIgnore
+     * @throws \Symfony\Component\Filesystem\Exception\IOException
+     * @throws \Symfony\Component\Finder\Exception\DirectoryNotFoundException
      */
     public static function tearDownAfterClass(): void
     {
         parent::tearDownAfterClass();
 
         if (LOGS !== TMP) {
-            /** @noinspection PhpUnhandledExceptionInspection */
-            Filesystem::instance()->unlinkRecursive(LOGS, ['.gitkeep', 'empty'], true);
+            Filesystem::unlinkRecursive(LOGS, ['.gitkeep', 'empty'], true);
         }
     }
 
@@ -156,9 +158,9 @@ abstract class TestCase extends CakeTestCase
      */
     protected function getLogFullPath(string $filename): string
     {
-        $filename .= Filesystem::instance()->getExtension($filename) ? '' : '.log';
+        $filename .= Filesystem::getExtension($filename) ? '' : '.log';
 
-        return Filesystem::instance()->makePathAbsolute($filename, LOGS);
+        return Filesystem::makePathAbsolute($filename, LOGS);
     }
 
     /**